requirements.txt
python-telegram-bot>=22.0
python-dotenv>=1.1.0
fastapi==0.115.11
uvicorn==0.22.0
pydantic==2.10.6
pydantic-settings==2.9.1
debugpy==1.8.14
langgraph==0.4.8
langchain-openai==0.3.19